+++ tests/xso/test_model.py
@@ -9,6 +9,8 @@ import unittest.mock
 
 import lxml.sax
 
+import multidict
+
 import aioxmpp.structs as structs
 import aioxmpp.xso as xso
 import aioxmpp.xso.model as xso_model
@@ -4930,6 +4932,209 @@ class TestChildValueMap(unittest.TestCase):
                 1: "a",
                 2: "b"
             }
+        )
+
+    def tearDown(self):
+        del self.Cls
+
+
+class TestChildValueMultiMap(unittest.TestCase):
+    class ChildXSO(xso.XSO):
+        TAG = ("uri:foo", "foo")
+
+        key = xso.Attr("k", type_=xso.Integer())
+        value = xso.Attr("v", type_=xso.Integer())
+
+    class ChildValueType(xso.AbstractType):
+        @classmethod
+        def get_formatted_type(cls):
+            return TestChildValueMultiMap.ChildXSO
+
+        def format(self, item):
+            key, value = item
+            item = TestChildValueMultiMap.ChildXSO()
+            item.key = key
+            item.value = value
+            return item
+
+        def parse(self, item):
+            return item.key, item.value
+
+    def setUp(self):
+        class Cls(xso.XSO):
+            TAG = ("uri:foo", "p")
+
+            values = xso_model.ChildValueMultiMap(
+                self.ChildValueType
+            )
+
+        self.Cls = Cls
+
+    def test_is_child_prop_base(self):
+        self.assertTrue(issubclass(
+            xso_model.ChildValueMultiMap,
+            xso_model._ChildPropBase
+        ))
+
+    def test_get_tag_map_and_classes(self):
+        self.assertDictEqual(
+            self.Cls.values.get_tag_map(),
+            {
+                self.ChildXSO.TAG: self.ChildXSO,
+            }
+        )
+        self.assertSetEqual(
+            self.Cls.values._classes,
+            {self.ChildXSO}
+        )
+
+    def test_registered_at_class_child_map(self):
+        self.assertIn(
+            self.ChildXSO.TAG,
+            self.Cls.CHILD_MAP
+        )
+        self.assertIs(
+            self.Cls.CHILD_MAP[self.ChildXSO.TAG],
+            self.Cls.values
+        )
+        self.assertIn(
+            self.Cls.values,
+            self.Cls.CHILD_PROPS
+        )
+
+    def test_get_on_class_returns_descriptor(self):
+        desc = xso_model.ChildValueList(self.ChildValueType)
+
+        class Cls(xso.XSO):
+            values = desc
+
+        self.assertIs(desc, Cls.values)
+
+    def test_initial_value_is_a_mutable_mapping(self):
+        obj = self.Cls()
+        self.assertIsInstance(
+            obj.values,
+            multidict.MultiDict)
+        self.assertEqual(obj.values, {})
+
+    def test_mapping_type(self):
+        class Cls(xso.XSO):
+            TAG = ("uri:foo", "p")
+
+            values = xso_model.ChildValueMultiMap(
+                self.ChildValueType,
+                mapping_type=multidict.CIMultiDict
+            )
+
+        obj = Cls()
+        self.assertIsInstance(
+            obj.values,
+            multidict.CIMultiDict)
+        self.assertEqual(obj.values, {})
+
+    def test_not_settable(self):
+        obj = self.Cls()
+
+        with self.assertRaises(AttributeError):
+            obj.values = {}
+
+    def test_from_events_uses__process_and_parse(self):
+        obj = self.Cls()
+
+        def process(mock, *args, **kwargs):
+            return mock(*args, **kwargs)
+            yield None
+
+        base = unittest.mock.Mock()
+        process_mock = base.process
+        base.process = functools.partial(process, process_mock)
+
+        with contextlib.ExitStack() as stack:
+            stack.enter_context(unittest.mock.patch.object(
+                self.ChildValueType,
+                "parse",
+                new=base.parse
+            ))
+
+            stack.enter_context(unittest.mock.patch.object(
+                self.Cls.values,
+                "_process",
+                new=base.process
+            ))
+
+            base.parse.return_value = ("x", "a")
+            with self.assertRaises(StopIteration):
+                gen = self.Cls.values.from_events(
+                    obj,
+                    base.ev_args,
+                    base.ctx,
+                )
+                next(gen)
+
+            base.parse.return_value = ("x", "b")
+            with self.assertRaises(StopIteration):
+                gen = self.Cls.values.from_events(
+                    obj,
+                    base.ev_args,
+                    base.ctx,
+                )
+                next(gen)
+
+            base.parse.return_value = ("z", "c")
+            with self.assertRaises(StopIteration):
+                gen = self.Cls.values.from_events(
+                    obj,
+                    base.ev_args,
+                    base.ctx,
+                )
+                next(gen)
+
+        calls = list(base.mock_calls)
+        self.assertSequenceEqual(
+            calls,
+            [
+                unittest.mock.call.process(obj, base.ev_args, base.ctx),
+                unittest.mock.call.parse(process_mock()),
+                unittest.mock.call.process(obj, base.ev_args, base.ctx),
+                unittest.mock.call.parse(process_mock()),
+                unittest.mock.call.process(obj, base.ev_args, base.ctx),
+                unittest.mock.call.parse(process_mock()),
+            ]
+        )
+
+        self.assertEqual(
+            obj.values,
+            multidict.MultiDict([
+                ("x", "a"),
+                ("x", "b"),
+                ("z", "c"),
+            ])
+        )
+
+    def test_to_sax_calls_format(self):
+        obj = self.Cls()
+        obj.values.add("x", "y")
+        obj.values.add("x", "z")
+
+        base = unittest.mock.Mock()
+
+        with contextlib.ExitStack() as stack:
+            stack.enter_context(unittest.mock.patch.object(
+                self.ChildValueType,
+                "format",
+                new=base.format
+            ))
+
+            self.Cls.values.to_sax(obj, base.dest)
+
+        self.assertSequenceEqual(
+            base.mock_calls,
+            [
+                unittest.mock.call.format(("x", "y")),
+                unittest.mock.call.format().unparse_to_sax(base.dest),
+                unittest.mock.call.format(("x", "z")),
+                unittest.mock.call.format().unparse_to_sax(base.dest),
+            ]
         )
 
     def tearDown(self):